adressing comments from Jan
