+++ workspaces/scorecard/packages/app-legacy/e2e-tests/scorecard.test.ts
@@ -86,7 +86,6 @@ import {
 import { installWebpackDevOverlayGuards } from './utils/devOverlays';
 
 test.describe('Scorecard Plugin Tests', () => {
-  const isNfs = process.env.APP_MODE === 'nfs';
   let page: Page;
   let catalogPage: CatalogPage;
   let scorecardPage: ScorecardPage;
@@ -318,6 +317,11 @@ test.describe('Scorecard Plugin Tests', () => {
   });
 
   test.describe('SonarQube Entity Scorecards', () => {
+    test.skip(
+      process.env.APP_MODE === 'nfs',
+      'Legacy-only: NFS uses grouped metric cards',
+    );
+
     test('Verify all SonarQube metrics display correctly', async ({}, testInfo) => {
       const sonarqubeMetricTitles = Object.entries(translations.metric)
         .filter(([key]) => key.startsWith('sonarqube.'))
@@ -329,38 +333,10 @@ test.describe('Scorecard Plugin Tests', () => {
       await catalogPage.openComponent('sonarqube-scorecard-only');
       await page.getByText('Scorecard', { exact: true }).click();
 
-      if (isNfs) {
-        const groupedMetricIds = new Set([
-          'sonarqube.qualityGate',
-          'sonarqube.codeCoverage',
-          'sonarqube.maintainabilityRating',
-          'sonarqube.openIssues',
-          'sonarqube.securityHotspots',
-          'sonarqube.securityRating',
-        ]);
-
-        const ungroupedMetricTitles = Object.entries(translations.metric)
-          .filter(
-            ([key]) =>
-              key.startsWith('sonarqube.') && !groupedMetricIds.has(key),
-          )
-          .map(([, value]) => (value as { title: string }).title);
-
-        for (const title of ungroupedMetricTitles) {
-          await expect(
-            page.getByText(title, { exact: true }).first(),
-          ).toBeVisible({ timeout: 10000 });
-        }
-
-        await expect(page.getByLabel('Security Vulnerabilities')).toBeVisible();
-        await expect(page.getByLabel('Code Quality')).toBeVisible();
-        await expect(page.getByLabel('SonarQube Coverage')).toBeVisible();
-      } else {
-        for (const title of sonarqubeMetricTitles) {
-          await expect(
-            page.getByText(title, { exact: true }).first(),
-          ).toBeVisible({ timeout: 10000 });
-        }
+      for (const title of sonarqubeMetricTitles) {
+        await expect(
+          page.getByText(title, { exact: true }).first(),
+        ).toBeVisible({ timeout: 10000 });
       }
 
       await runAccessibilityTests(page, testInfo);
@@ -396,37 +372,33 @@ test.describe('Scorecard Plugin Tests', () => {
         await expect(card).toContainText(value);
       }
 
-      if (isNfs) {
-        await expect(page.getByLabel('Code Quality')).toBeVisible();
-      } else {
-        const legacyOnlyValues: Record<string, string> = {
-          [translations.metric['sonarqube.qualityGate'].title]:
-            translations.thresholds.success,
-          [translations.metric['sonarqube.openIssues'].title]: '3',
-          [translations.metric['sonarqube.securityRating'].title]: '1',
-          [translations.metric['sonarqube.securityHotspots'].title]: '2',
-          [translations.metric['sonarqube.maintainabilityRating'].title]: '1',
-          [translations.metric['sonarqube.codeCoverage'].title]: '82.5',
-        };
+      const legacyOnlyValues: Record<string, string> = {
+        [translations.metric['sonarqube.qualityGate'].title]:
+          translations.thresholds.success,
+        [translations.metric['sonarqube.openIssues'].title]: '3',
+        [translations.metric['sonarqube.securityRating'].title]: '1',
+        [translations.metric['sonarqube.securityHotspots'].title]: '2',
+        [translations.metric['sonarqube.maintainabilityRating'].title]: '1',
+        [translations.metric['sonarqube.codeCoverage'].title]: '82.5',
+      };
 
-        for (const [title, value] of Object.entries(legacyOnlyValues)) {
-          const card = page
-            .locator('[role="article"]')
-            .filter({ hasText: title })
-            .first();
-          await expect(card).toContainText(value);
-        }
-
-        const qualityGateCard = page
+      for (const [title, value] of Object.entries(legacyOnlyValues)) {
+        const card = page
           .locator('[role="article"]')
-          .filter({
-            hasText: translations.metric['sonarqube.qualityGate'].title,
-          })
+          .filter({ hasText: title })
           .first();
-        await expect(
-          qualityGateCard.getByTestId('CheckCircleOutlineIcon'),
-        ).toBeVisible();
+        await expect(card).toContainText(value);
       }
+
+      const qualityGateCard = page
+        .locator('[role="article"]')
+        .filter({
+          hasText: translations.metric['sonarqube.qualityGate'].title,
+        })
+        .first();
+      await expect(
+        qualityGateCard.getByTestId('CheckCircleOutlineIcon'),
+      ).toBeVisible();
     });
 
     test('Verify SonarQube quality gate failure state', async () => {
@@ -439,29 +411,16 @@ test.describe('Scorecard Plugin Tests', () => {
       await catalogPage.openComponent('sonarqube-scorecard-only');
       await page.getByText('Scorecard', { exact: true }).click();
 
-      if (isNfs) {
-        const codeQualityCard = page
-          .locator('[role="article"]')
-          .filter({ hasText: 'Code Quality' })
-          .first();
-        await expect(codeQualityCard).toBeVisible({ timeout: 10000 });
-
-        const errorBucket = codeQualityCard
-          .locator('[role="button"]')
-          .filter({ hasText: translations.thresholds.error });
-        await expect(errorBucket).toBeVisible();
-      } else {
-        const qualityGateCard = page
-          .locator('[role="article"]')
-          .filter({
-            hasText: translations.metric['sonarqube.qualityGate'].title,
-          })
-          .first();
-        await expect(qualityGateCard).toBeVisible({ timeout: 10000 });
-        await expect(
-          qualityGateCard.getByTestId('DangerousOutlinedIcon'),
-        ).toBeVisible();
-      }
+      const qualityGateCard = page
+        .locator('[role="article"]')
+        .filter({
+          hasText: translations.metric['sonarqube.qualityGate'].title,
+        })
+        .first();
+      await expect(qualityGateCard).toBeVisible({ timeout: 10000 });
+      await expect(
+        qualityGateCard.getByTestId('DangerousOutlinedIcon'),
+      ).toBeVisible();
     });
 
     test('Verify empty state for sonarqube entity with no metrics', async () => {
@@ -1145,152 +1104,6 @@ test.describe('Scorecard Plugin Tests', () => {
         );
         await expect(card).toContainText('customUnknownAggregationKind');
       });
-    });
-  });
-
-  test.describe('Metric Group Cards', () => {
-    test.skip(!isNfs, 'MetricGroupCard is only available in NFS mode');
-
-    test.beforeAll(async () => {
-      await mockSonarqubeScorecardResponse(page, sonarqubeScorecardResponse);
-      await catalogPage.openCatalog();
-      await catalogPage.openComponent('sonarqube-scorecard-only');
-      await page.getByText('Scorecard', { exact: true }).click();
-      await expect(page.getByLabel('Security Vulnerabilities')).toBeVisible({
-        timeout: 15000,
-      });
-    });
-
-    test('Verify group cards render with titles, descriptions, and bucket tiles', async ({}, testInfo) => {
-      // Security Vulnerabilities: sonarqube.securityRating(success) = 1 success
-      const securityCard = scorecardPage.getGroupCard(
-        'Security Vulnerabilities',
-      );
-      await expect(page.getByLabel('Security Vulnerabilities')).toBeVisible();
-      await expect(
-        securityCard.getByText(
-          'Track security issues across your repositories',
-        ),
-      ).toBeVisible();
-      await expect(
-        scorecardPage.getBucketTile(securityCard, 'success'),
-      ).toContainText('1');
-
-      // Code Quality: qualityGate(success) + codeCoverage(success) + maintainabilityRating(success) = 3 success
-      const codeQualityCard = scorecardPage.getGroupCard('Code Quality');
-      await expect(page.getByLabel('Code Quality')).toBeVisible();
-      await expect(
-        codeQualityCard.getByText('Code quality and maintainability metrics'),
-      ).toBeVisible();
-      await expect(
-        scorecardPage.getBucketTile(codeQualityCard, 'success'),
-      ).toContainText('3');
-
-      // SonarQube Coverage: openIssues(warning) + securityHotspots(warning) + securityRating(success)
-      const coverageCard = scorecardPage.getGroupCard('SonarQube Coverage');
-      await expect(page.getByLabel('SonarQube Coverage')).toBeVisible();
-      await expect(
-        coverageCard.getByText('SonarQube coverage metrics'),
-      ).toBeVisible();
-      await expect(
-        scorecardPage.getBucketTile(coverageCard, 'warning'),
-      ).toContainText('2');
-      await expect(
-        scorecardPage.getBucketTile(coverageCard, 'success'),
-      ).toContainText('1');
-
-      await runAccessibilityTests(page, testInfo);
-    });
-
-    test('Verify data sources dialog opens from menu', async ({}, testInfo) => {
-      const securityCard = scorecardPage.getGroupCard(
-        'Security Vulnerabilities',
-      );
-      const dialog = await scorecardPage.openDataSourcesDialog(securityCard);
-
-      await expect(dialog).toContainText(
-        scorecardPage.getDialogTitle('Security Vulnerabilities'),
-      );
-
-      await expect(
-        dialog.getByText(translations.dataSourcesDialog.columns.plugin),
-      ).toBeVisible();
-      await expect(
-        dialog.getByText(translations.dataSourcesDialog.columns.check),
-      ).toBeVisible();
-      await expect(
-        dialog.getByText(translations.dataSourcesDialog.columns.value, {
-          exact: true,
-        }),
-      ).toBeVisible();
-      await expect(
-        dialog.getByText(translations.dataSourcesDialog.columns.status),
-      ).toBeVisible();
-      await expect(
-        dialog.getByText(translations.dataSourcesDialog.columns.lastSynced),
-      ).toBeVisible();
-
-      await runAccessibilityTests(page, testInfo, undefined, {
-        disableRules: ['color-contrast'],
-      });
-
-      await scorecardPage.closeDialog(dialog);
-      await expect(dialog).not.toBeVisible();
-    });
-
-    test('Verify filter pills filter data sources by threshold in SonarQube Coverage', async () => {
-      const coverageCard = scorecardPage.getGroupCard('SonarQube Coverage');
-      const dialog = await scorecardPage.openDataSourcesDialog(coverageCard);
-      const tableRows = scorecardPage.getTableRows(dialog);
-
-      // SonarQube Coverage has 3 metrics: openIssues(warning), securityHotspots(warning), securityRating(success)
-      await expect(tableRows).toHaveCount(3);
-
-      // Click "Warning" filter → should show only the 2 warning metrics
-      const warningPill = scorecardPage.getFilterPill(dialog, 'warning');
-      await warningPill.click();
-      await expect(warningPill).toHaveAttribute('aria-pressed', 'true');
-      await expect(tableRows).toHaveCount(2);
-      for (const row of await tableRows.all()) {
-        await expect(row).toContainText(translations.thresholds.warning);
-      }
-
-      // Deactivate "Warning" → all 3 rows visible again
-      await warningPill.click();
-      await expect(warningPill).toHaveAttribute('aria-pressed', 'false');
-      await expect(tableRows).toHaveCount(3);
-
-      // Click "Success" filter → should show only 1 success metric
-      const successPill = scorecardPage.getFilterPill(dialog, 'success');
-      await successPill.click();
-      await expect(successPill).toHaveAttribute('aria-pressed', 'true');
-      await expect(tableRows).toHaveCount(1);
-      for (const row of await tableRows.all()) {
-        await expect(row).toContainText(translations.thresholds.success);
-      }
-
-      await scorecardPage.closeDialog(dialog);
-    });
-
-    test('Verify clicking bucket tile opens dialog with filter pre-applied', async () => {
-      const coverageCard = scorecardPage.getGroupCard('SonarQube Coverage');
-
-      // Click the "Warning" bucket tile (count = 2)
-      await scorecardPage.getBucketTile(coverageCard, 'warning').click();
-
-      const dialog = page.locator('[role="dialog"]');
-      await expect(dialog).toBeVisible({ timeout: 5000 });
-
-      // Filter is pre-applied from the tile click
-      const activePill = dialog.locator('[role="button"][aria-pressed="true"]');
-      await expect(activePill).toBeVisible();
-      const tableRows = scorecardPage.getTableRows(dialog);
-      await expect(tableRows).toHaveCount(2);
-      for (const row of await tableRows.all()) {
-        await expect(row).toContainText(translations.thresholds.warning);
-      }
-
-      await scorecardPage.closeDialog(dialog);
     });
   });
 });
